task/vector_and_matrices/requirements.txt
numpy==1.26.4
python-dotenv==1.0.1
